Skipped also the paywall tests
