+++ examples/webbilling-demo/src/tests/constants.ts
@@ -1,0 +1,5 @@
+export const CARD_SELECTOR = "div.card";
+export const PACKAGE_SELECTOR = "button.rc-pw-package";
+export const RC_PAYWALL_TEST_OFFERING_ID = "rc_paywalls_e2e_test_2";
+export const RC_PAYWALL_TEST_OFFERING_ID_WITH_VARIABLES =
+  "rc_paywalls_e2e_test_variables_2";
